split code to individual classes
